Move the pre-commit tests in a separate file
